added txt file to include offensive and defensive statistic tokens, also moved playergoalcontribution

+++ soccer_data_program/src/statistics_models/DefensiveStat.java
@@ -1,5 +1,5 @@
 package statistics_models;
 
 public abstract class DefensiveStat {
-  public abstract int calculate();
+  public abstract double calculate();
 }
